pyproject.toml
[project]
name = "optimql"
version = "0.1.0"
description = "Autonomous Multi-Agent AI Database Optimizer"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.20.0",
    "requests>=2.28.0",
]

[tool.pyright]
pythonVersion = "3.10"
venvPath = "."
venv = "venv"
extraPaths = ["."]
